requirements.txt
torch>=2.0.0
transformers>=4.35.0
trl>=0.7.0
peft>=0.7.0
datasets>=2.14.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
pyyaml>=6.0
wandb>=0.15.0
